% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{get_cluster_centroids}
\alias{get_cluster_centroids}
\title{Determine the center of each cluster in a biplot}
\usage{
get_cluster_centroids(clusters, scores)
}
\arguments{
\item{clusters}{A vector containing the cluster labels for all 
observations}

\item{scores}{A data frame containing the PCA scores.}
}
\description{
This calculates the x and y coordinates for the center of each cluster
for data that has been subjected to dimensionality reduction followed by
clustering.
}
