% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{influential_species}
\alias{influential_species}
\title{Get the most influential species}
\usage{
influential_species(cov_table)
}
\description{
Get the top 5 species ranked by the amount of variance.
}
\details{
@param cov_table The return value of project_variables
 @value A covariance table
 @export
}
