% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{optimize_hclust}
\alias{optimize_hclust}
\title{Determine the number of clusters in hierarchical clustering}
\usage{
optimize_hclust(matrix, tree)
}
\arguments{
\item{matrix}{A (dis) similarity matrix}

\item{tree}{A hclust object}
}
\description{
Find the optimum number of clusters for the hclust algorithm based on the
maximized Calinski-Harabasz index. Note that this tests clusters between
2 and 20.
}
