% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{optimize_kmeans}
\alias{optimize_kmeans}
\title{Determine k in the kmeans algorithm}
\usage{
optimize_kmeans(data)
}
\arguments{
\item{data}{Data to be passed to kmeans}
}
\value{
a data frame consisting of the number of clusters and total WSS
}
\description{
Find the optimum number of clusters for the kmeans algorithm based on 
total within sum of squared error. Note that this tests clusters
between 1 and 20.
}
