% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{project_variables}
\alias{project_variables}
\title{Project the original data on a biplot}
\usage{
project_variables(pcoa, original_data)
}
\arguments{
\item{pcoa}{An object of class pcoa}

\item{original_data}{The data used to generate the object passed as pcoa}
}
\description{
Project the original data on to the new Principal Coordinates. This
can be used to generate arrows similar to what is observed on a biplot.
Based on the implementation in the ape package.
}
